+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog for *TopoBank*
 
-## 1.0.0 (not published yet)
+## 0.12.0 (2021-05-20)
+
+- use acquisition time from data file as initial 
+  value for 'measurement date' after upload, if present (#433)
+- added buttons for selection and deselection of
+  measurements and averages to analysis result plots (#623)
+- on viewing a measurement, loading of the plot is
+  now done in the background for faster page access (#597)
+- fixes internal server error because of failing thumbnail
+  generation (#662)
+- fixes internal server error display, now show a more
+  friendly page with contact link and button to go 
+  back to site (#666)
+- fixes server error when downloading analyses data
+  for PSD for surfaces with only one measurement (#664)
 
 ## 0.11.1 (2021-05-11)